+++ src/app/components/penyelenggara-home/penyelenggara-home.component.ts
@@ -5,6 +5,9 @@ import { MatSort } from '@angular/material/sort';
 import { MatTableDataSource } from '@angular/material/table';
 import { Router } from '@angular/router';
 import { TokenStorageService } from 'src/app/_services/token-storage.service';
+import { SpinnerService } from '../../_services/spinner.service';
+import Swal from 'sweetalert2';
+import { faPencilAlt, faTrash } from '@fortawesome/free-solid-svg-icons';
 
 // created by jason
 export interface Organizer {
@@ -34,6 +37,9 @@ const DELETE_ORGANIZER_API = 'http://10.1.137.50:8762/delete/'
 })
 export class PenyelenggaraHomeComponent implements OnInit {
 
+  faEdit = faPencilAlt;
+  faDelete = faTrash;
+
   // token for get anything data
   httpOptions_base = {
     headers: new HttpHeaders().set(
@@ -48,14 +54,16 @@ export class PenyelenggaraHomeComponent implements OnInit {
   @ViewChild(MatPaginator) paginator!: MatPaginator;
   @ViewChild(MatSort) sort!: MatSort;
 
-  //DI 
+  //DI
   constructor(
     private http: HttpClient,
     private router: Router,
-    private token: TokenStorageService
+    private token: TokenStorageService,
+    private spinner: SpinnerService
   ) { }
 
   ngOnInit(): void {
+    this.spinner.isLoading = true
     this.getAllData()
   }
 
@@ -74,25 +82,51 @@ export class PenyelenggaraHomeComponent implements OnInit {
         this.dataSource = new MatTableDataSource(isi.content)
         this.dataSource.paginator = this.paginator
         this.dataSource.sort = this.sort
+        this.spinner.isLoading = false
       },
         err => {
-          alert("Error fetching data")
-         
+          this.spinner.isLoading = false
+          Swal.fire({
+            icon: 'error',
+            title: 'Oops...',
+            text: err.error.message,
+          })
         }
       )
   }
 
   deleteOrganizer(id: string) {
-    this.http.delete<any>(DELETE_ORGANIZER_API + id, this.httpOptions_base)
-      .subscribe(isi => {
-       
-        this.getAllData()
-      },
-        err => {
-          alert("Error while deleting")
-          
-        }
-      )
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.value) {
+        this.http.delete<any>(DELETE_ORGANIZER_API + id, this.httpOptions_base)
+        .subscribe(isi => {
+          this.getAllData()
+        },
+          err => {
+            Swal.fire({
+              icon: 'error',
+              title: 'Oops...',
+              text: err.error.message,
+            })
+          }
+        )
+      } else {
+        Swal.fire(
+          'Cancelled',
+          'Your seller file is safe :)',
+          'error'
+        )
+      }
+    })
+
   }
 
 }
